[enclave] extrated io out of utils
