+++ enclave/src/state.rs
@@ -5,10 +5,10 @@ use sgx_types::*;
 use log::*;
 
 use crate::aes;
-use crate::utils::*;
+use crate::io;
 
 pub fn read(path: &str) -> SgxResult<Vec<u8>> {
-	let mut bytes = match read_plaintext(path) {
+	let mut bytes = match io::read_plaintext(path) {
 		Ok(vec) => match vec.len() {
 			0 => return Ok(vec),
 			_ => vec,
@@ -27,7 +27,7 @@ pub fn write_encrypted(bytes: &mut Vec<u8>, path: &str) -> SgxResult<sgx_status_
 
 	aes::de_or_encrypt(bytes)?;
 
-	write_plaintext(&bytes, path)?;
+	io::write_plaintext(&bytes, path)?;
 	Ok(sgx_status_t::SGX_SUCCESS)
 }
